backend/requirements.txt
fastapi[standard]
azure-identity
python-dotenv
semantic-kernel